backend/foodgram/requirements.txt
asgiref==3.6.0
Django==4.2.1
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djoser==2.2.0
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sqlparse==0.4.4
urllib3==2.0.2
webcolors==1.13
